# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Films API",
        "version": "0.1"
    },
    "paths": {
        "/api/signup": {
            "post": {
                "tags": [
                    "Authentification"
                ],
                "summary": "Creates a new user and token",
                "description": "Maximum of 5 calls per minute",
                "operationId": "65b7ef1b1e076c033bf91055f53bd9a1",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "login": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    },
                                    "password_confirmation": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "last_name": {
                                        "type": "string"
                                    },
                                    "first_name": {
                                        "type": "string"
                                    },
                                    "role_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "422": {
                        "description": "Invalid data"
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            }
        },
        "/api/signin": {
            "post": {
                "tags": [
                    "Authentification"
                ],
                "summary": "Connect to an user in bd and creates a token",
                "description": "Maximum of 5 calls per minute",
                "operationId": "b5cee56757c3c79c5b5f115fdc33f460",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "422": {
                        "description": "Invalid data"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            }
        },
        "/api/signout": {
            "post": {
                "tags": [
                    "Authentification"
                ],
                "summary": "Deletes all existing tokens for user",
                "description": "Maximum of 5 calls per minute",
                "operationId": "b46f8611fc7d1123f2a3d573cee661c4",
                "responses": {
                    "204": {
                        "description": "No content"
                    }
                }
            }
        }
    },
    "components": {
        "responses": {
            "500": {
                "description": "Server error"
            }
        }
    }
}